requirements.txt
django==3.2.6
python-dotenv==0.19.0
djangorestframework==3.12.4
soupsieve==2.2.1
requests==2.26.0